refactor: calculator-preview -> extracted schema into separate file
