+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import { GetStaticProps } from "next";
 import { useTranslations } from "next-intl";
 import Accordions from "@/components/homepage/homepage-accordions";
-import CalculatorPreview from "@/components/homepage/calculator-preview";
+import CalculatorPreview from "@/components/homepage/calculator-preview/calculator-preview";
 import LangButton from "@/components/unique/lang-button";
 
 const Homepage = () => {
